type checking for arrays and pointer math

+++ list.c
@@ -77,9 +77,9 @@ void list_append(List *list, ListNode *head)
 //
 // Return the number of elements in a list
 //
-int list_count(List *list)
+size_t list_count(List *list)
 {
-    int count = 0;
+    size_t count = 0;
     ListNode *node;
 
     for (node = list->head; node; node = node->next) {
